Conflicts:
	.travis.yml
	ChangeLog
	README
	familyconnections/gallery/index.php

+++ familyconnections/admin/members.php
@@ -18,7 +18,7 @@ define('GALLERY_PREFIX', '../gallery/');
 
 require URL_PREFIX.'fcms.php';
 
-load('admin_members');
+load('admin_members', 'phpass');
 
 init('admin/');
 
@@ -357,7 +357,12 @@ Event.observe(window, \'load\', function() {
         echo sprintf(T_pgettext('%s is a name of a table that gets updated.', 'Update [%s] complete.'), 'fcms_address').'<br/>';
 
         // Update all occurences of merge id with id
-        $this->fcmsAdminMembers->mergeMember($id, $merge);
+        if (!$this->fcmsAdminMembers->mergeMember($id, $merge))
+        {
+            $this->fcmsError->displayError();
+            $this->displayFooter();
+            return;
+        }
 
         // Delete merge id
         $sql = "DELETE FROM `fcms_users`
@@ -451,12 +456,14 @@ Event.observe(window, \'load\', function() {
         $sex      = strip_tags($_POST['sex']);
         $email    = strip_tags($_POST['email']);
         $username = strip_tags($_POST['username']);
-        $md5pass  = md5($_POST['password']);
+
+        $hasher       = new PasswordHash(8, FALSE);
+        $hashPassword = $hasher->HashPassword($_POST['password']);
 
         // Create new member
         $sql = "INSERT INTO `fcms_users`
                     (`access`, `joindate`, `fname`, `mname`, `lname`, `maiden`, `sex`, `email`, `dob_year`, `dob_month`, `dob_day`,
-                        `username`, `password`, `activated`)
+                        `username`, `phpass`, `activated`)
                 VALUES
                     (3, NOW(), ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 1)";
 
@@ -471,7 +478,7 @@ Event.observe(window, \'load\', function() {
             $month, 
             $day,
             $username, 
-            $md5pass, 
+            $hashPassword, 
         );
 
         $lastid = $this->fcmsDatabase->insert($sql, $params);
@@ -672,8 +679,10 @@ Event.observe(window, \'load\', function() {
 
         if ($_POST['password'])
         {
+            $hasher       = new PasswordHash(8, FALSE);
+
             $sql     .= "`password` = ?, ";
-            $params[] = md5($_POST['password']);
+            $params[] = $hasher->HashPassword($_POST['password']);
 
             $sitename = getSiteName();
             $subject  = getSiteName().': '.T_('Password Change');
